requirements.txt
openai==2.32.0
python-dotenv==1.2.2
python-telegram-bot==22.7
